change: update EWMA
